Cleanups for setup.el
Create some variables that can be set to customize dylan-lsp. These should be
defcustom eventually but I'm not going to do that until things settle down
more.

Don't expect the lsp-dylan binary to be in _build/bin; just make sure it's on
PATH. This way I can link it from ~/dylan/bin which is already on my path, and
I don't have to be in the lsp-dylan directory when I start emacs.