Add a skeleton of a test suite
Actual tests to be added later.

For the test suite we need to use lsp-dylan and not call the main function, so
I broke lsp-dylan into an lsp-dylan shared library plus a dylan-lsp-server
executable. The latter being a nicer name than lsp-dylan-app, IMO, without the
confusing of using "executable:" in the LID file.

While I was at it I broke lsp-dylan into interface and implementation modules.