Add a test for send-request
I saw the two lines below and thought "this must never be called because
rebinding id like that would be an error", but of course it's id-setter that's
being called in the second line so it's not an error. I'll probably rename id
to session-id on general principle, but in the meantime at least we have a
basic test.

```dylan
  let id = session.id;
  session.id := id + 1;
```