Changes to textDocument/didOpen
Moves much of what was in test-open-project to the textDocument/didOpen handler
and removes test-open-project. This means the project is no longer opened in
the "initialize" handler.

This now REQUIRES that the opened file be in a workspace directory with a
"default-library" setting in workspace.json. The plan is to relax that
requirement somewhat in the future by implementing code to find "root"
libraries in the workspace.