Fix server command-line options
Because of a bug in command-line-parser there's no way to turn off a
boolean command-line flag that defaults to true, so I changed both debug flags
to default to false and changed the elisp to have one variable that holds a
list of command-line options instead. Seems a little more flexible, too.

There should be no change in default behavior since both options are included
on the command-line by default.