Change how projects are opened and tracked
There are a bunch of changes here. I should've done a better job separating
this into several changes, but I don't think it's worth the effort now.

When any file is opened (textDocument/didOpen) a `<document>` is created if not
already open and the project (a `<project-object>`) is stored in that
document. The globals for tracking the open project/library/module have been
removed and we always go through the document now.

`find-open-project` can be used to search all open documents to find a
project (library) with a given name.

`open-lsp-project` (formerly `test-open-project`) uses the
`workspaces:dylan-tool` module to figure out which library a file belongs to
and re-uses the open project for that library, if any, or opens a new project
and stores it in the document that is being opened.

This seems to work well for me, but for it to work in used libraries (i.e., the
dependencies of the first project opened) it requires that
`OPEN_DYLAN_USER_ROOT` be set so that the build products will be found. For
example, if I start by opening lsp-dylan.dylan the `lsp-dylan` project is
opened. If I then `M-. file-module` I end up in OD code. For LSP to work
in *that* file `O_D_U_R` must be set.

I think we can remove the requirement for `OPEN_DYLAN_USER_ROOT` to be
set (which I consider to be error prone and unintuitive) by inferring it when
the first file is opened. Either way, we need to document how that works.
Future work....